Transaction page restructuring.
