+++ src/App.jsx
@@ -2,7 +2,8 @@ import { Provider } from 'react-redux'
 import { store } from './app/store';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { Help, Profile, UserHome } from './pages/user';
-import { AdminHome, Transaction,Settings } from './pages/admin';
+import { AdminHome, Settings } from './pages/admin';
+import { Transaction } from './pages/admin';
 
 import './App.css'
