+++ src/pages/admin/index.js
@@ -1,5 +1,5 @@
 import { AdminHome } from './AdminHome';
-import { Transaction } from './Transaction';
+import { Transaction } from './transactions/Transaction';
 import { Settings } from './Settings';
 
 export { AdminHome, Transaction, Settings };
